move Countdown into its own component
